+++ subgraph/fluence-graph.sh
@@ -134,7 +134,8 @@ esac
 case "$action" in
   deploy)
     echo "Deploying subgraph on ${network} stand with subgraph name: $SUBGRAPH_NAME and version label $SUBGRAPH_VERSION_LABEL..."
-    graph deploy --node ${GRAPHNODE_URL} --ipfs ${IPFS_URL} --network ${SUBGRAPH_NETWORK} --network-file configs/${network}-networks-config.json --version-label ${SUBGRAPH_VERSION_LABEL} ${SUBGRAPH_NAME}
+    auth_header=$(echo -n ${BASIC_AUTH_SUBGRAPH} | base64)
+    graph deploy --node ${GRAPHNODE_URL} --headers "{\"Authorization\": \"Basic ${auth_header}\"}" --ipfs ${IPFS_URL} --network ${SUBGRAPH_NETWORK} --network-file configs/${network}-networks-config.json --version-label ${SUBGRAPH_VERSION_LABEL} ${SUBGRAPH_NAME} \
     ;;
   create)
     echo "Creating subgraph on Fluence with name: $SUBGRAPH_NAME..."