+++ subgraph/src/mappings/core.ts
@@ -1,4 +1,7 @@
-import { createOrLoadGraphNetwork } from "../models";
+import {
+  createOrLoadGraphNetwork,
+  createOrLoadProvider
+} from "../models";
 import {
   Initialized,
   WhitelistAccessGranted,
@@ -9,17 +12,21 @@ import {
   getEpochDuration,
   getInitTimestamp,
   getMinRequiredProofsPerEpoch,
+  getPrecision,
 } from "../contracts";
 import { Provider } from "../../generated/schema";
+import { formatAddress } from "./utils";
+import { log } from "@graphprotocol/graph-ts/index";
 
 export function handleInitialized(event: Initialized): void {
   let graphNetwork = createOrLoadGraphNetwork();
   graphNetwork.coreEpochDuration = getEpochDuration(event.address);
-  graphNetwork.coreContractAddress = event.address.toHexString();
+  graphNetwork.coreContractAddress = formatAddress(event.address);
   graphNetwork.initTimestamp = getInitTimestamp(event.address);
   graphNetwork.capacityMaxFailedRatio = getCapacityMaxFailedRatio(
     event.address,
   ).toI32();
+  graphNetwork.corePrecision = getPrecision(event.address).toI32();
   graphNetwork.minRequiredProofsPerEpoch = getMinRequiredProofsPerEpoch(
     event.address,
   ).toI32();
@@ -30,7 +37,7 @@ export function handleInitialized(event: Initialized): void {
 export function handleWhitelistAccessGranted(
   event: WhitelistAccessGranted,
 ): void {
-  let provider = Provider.load(event.params.account.toHexString()) as Provider;
+  let provider = createOrLoadProvider(formatAddress(event.params.account), event.block.timestamp);
   provider.approved = true;
   provider.save();
 }
@@ -38,7 +45,7 @@ export function handleWhitelistAccessGranted(
 export function handleWhitelistAccessRevoked(
   event: WhitelistAccessRevoked,
 ): void {
-  let provider = Provider.load(event.params.account.toHexString()) as Provider;
+  let provider = createOrLoadProvider(formatAddress(event.params.account), event.block.timestamp);
   provider.approved = false;
   provider.save();
 }